+++ templates/micro-main/src/query-entery-file.ts
