+++ templates/micro-main/src/tools/event-bus.ts
