+++ templates/micro-main/src/tools/global-state.ts
